requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
psycopg2-binary>=2.9.0
redis>=4.5.0
celery>=5.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pydantic-settings>=2.0.0
crewai
pyautogen>=0.2.0
numpy>=1.24.0
